Basic fixes for Windows #354

+++ codalab/lib/path_util.py
@@ -214,9 +214,9 @@ def get_size(path, dirs_and_files=None):
     Does not include symlinked files and directories.
     '''
     if os.path.islink(path) or not os.path.isdir(path):
-        return os.lstat(path).st_size
+        return long(os.lstat(path).st_size)
     dirs_and_files = dirs_and_files or recursive_ls(path)
-    return sum(os.lstat(path).st_size for path in itertools.chain(*dirs_and_files))
+    return sum(long(os.lstat(path).st_size) for path in itertools.chain(*dirs_and_files))
 
 
 def hash_directory(path, dirs_and_files=None):
